add retry request to github api on 202 response
https://developer.github.com/v3/repos/statistics/#commit-activity
“Computing repository statistics is an expensive operation, so we try
to return cached data whenever possible. If the data hasn’t been cached
when you query a repository’s statistics, you’ll receive a 202
response; a background job is also fired to start compiling these
statistics. Give the job a few moments to complete, and then submit the
request again. If the job has completed, that request will receive a
200 response with the statistics in the response body.”